+++ app/src/main/java/com/jerboa/ui/components/ban/BanFromCommunityActivity.kt
@@ -49,7 +49,7 @@ fun BanFromCommunityActivity(
 
     var removeData by rememberSaveable { mutableStateOf(false) }
     var permaBan by rememberSaveable { mutableStateOf(false) }
-    var expireDays: Int? by rememberSaveable { mutableStateOf(null) }
+    var expireDays: Long? by rememberSaveable { mutableStateOf(null) }
 
     var reason by rememberSaveable(stateSaver = TextFieldValue.Saver) {
         mutableStateOf(
